requirements.txt
beautifulsoup4==4.11.1
PyYAML==6.0
requests==2.28.1